# publicsuffix: spruce up code gen and speed up PublicSuffix